+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/exception/ValidationError.java
@@ -32,7 +32,15 @@ public class ValidationError implements Serializable {
 
   public static ValidationError forDeletionUsage(Deletable object, long usage, String pluralizedReferrer) {
     return new ValidationError(
-        String.format("%s %s is used by %d %s", object.getDeleteType(), object.getDeleteDescription(), usage, pluralizedReferrer));
+        String.format("%s '%s' is used by %d %s", object.getDeleteType(), object.getDeleteDescription(), usage, pluralizedReferrer));
+  }
+
+  public static ValidationError forDuplicate(String typeLabel, String property) {
+    return forDuplicate(typeLabel, property, property);
+  }
+
+  public static ValidationError forDuplicate(String typeLabel, String property, String propertyLabel) {
+    return new ValidationError(property, String.format("There is already a %s with this %s", typeLabel, propertyLabel));
   }
 
   /**