+++ miso-web/src/main/webapp/scripts/list_sequencingcontroltype.js
@@ -1,0 +1,28 @@
+ListTarget.sequencingcontroltype = {
+  name: "Sequencing Control Types",
+  getUserManualUrl: function() {
+    return Urls.external.userManual('type_data', 'sequencing-control-types');
+  },
+  createUrl: function(config, projectId) {
+    throw new Error("Must be provided statically");
+  },
+  getQueryUrl: null,
+  createBulkActions: function(config, projectId) {
+    var actions = HotTarget.sequencingcontroltype.getBulkActions(config);
+    if (config.isAdmin) {
+      actions.push(ListUtils.createBulkDeleteAction('Sequencing Control Types', 'sequencingcontroltypes', Utils.array.getAlias));
+    }
+    return actions;
+  },
+  createStaticActions: function(config, projectId) {
+    return config.isAdmin ? [ListUtils.createStaticAddAction('Sequencing Control Types', 'sequencingcontroltype')] : [];
+  },
+  createColumns: function(config, projectId) {
+    return [{
+      sTitle: "Alias",
+      mData: "alias",
+      iSortPriority: 1,
+      bSortDirection: true
+    }];
+  }
+};