+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateSequencingControlTypeDao.java
@@ -1,9 +1,12 @@
 package uk.ac.bbsrc.tgac.miso.persistence.impl;
 
+import java.io.IOException;
+
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import uk.ac.bbsrc.tgac.miso.core.data.SequencingControlType;
+import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleImpl;
 import uk.ac.bbsrc.tgac.miso.persistence.SequencingControlTypeDao;
 
 @Repository
@@ -12,6 +15,16 @@ public class HibernateSequencingControlTypeDao extends HibernateSaveDao<Sequenci
 
   public HibernateSequencingControlTypeDao() {
     super(SequencingControlType.class);
+  }
+
+  @Override
+  public SequencingControlType getByAlias(String alias) throws IOException {
+    return getBy("alias", alias);
+  }
+
+  @Override
+  public long getUsage(SequencingControlType sequencingControlType) throws IOException {
+    return getUsageBy(SampleImpl.class, "sequencingControlType", sequencingControlType);
   }
 
 }